+++ src/renderer/components/chat/messages/hooks/useMessageSelectionController.ts
@@ -10,8 +10,8 @@ import {
   getSelectedMessagesPlainText,
   getSelectedMessagesRichClipboardContent
 } from '@renderer/components/chat/messages/utils/messageSelection'
+import { messagesToMarkdown } from '@renderer/services/ExportService'
 import { formatErrorMessageWithPrefix } from '@renderer/utils/error'
-import { messagesToMarkdown } from '@renderer/utils/export'
 import type { CherryMessagePart } from '@shared/data/types/message'
 import { useCallback, useEffect, useMemo } from 'react'
 import { useTranslation } from 'react-i18next'
