+++ src/renderer/pages/home/Tabs/components/__tests__/Topics.test.tsx
@@ -206,8 +206,7 @@ vi.mock('@renderer/components/Popups/SaveToKnowledgePopup', () => ({
   default: { showForTopic: vi.fn() }
 }))
 
-vi.mock('@renderer/utils/export', () => ({
-  copyTopicAsMarkdown: vi.fn(),
+vi.mock('@renderer/services/ExportService', () => ({
   exportMarkdownToJoplin: vi.fn(),
   exportMarkdownToSiyuan: vi.fn(),
   exportMarkdownToYuque: vi.fn(),
@@ -217,7 +216,7 @@ vi.mock('@renderer/utils/export', () => ({
   topicToMarkdown: vi.fn().mockResolvedValue('# topic')
 }))
 
-vi.mock('@renderer/utils/copy', () => ({
+vi.mock('@renderer/services/CopyService', () => ({
   copyTopicAsMarkdown: vi.fn(),
   copyTopicAsPlainText: vi.fn()
 }))
